chore: incorporate cdk-assets-schema into cloud-assembly-schema
assets.json is part of the cloud assembly and as such we want to make sure it's schema and compatibility is managed together with the cloud assembly manifest.

To do that, we extended the cloud assembly schema "root" to include two keys `manifest` and `assets` and incorporated all the types from cdk-assets-schema into the JSON schema.

Now we can leverage the compatibility checks and versioning from cloud-assembly-schema instead of the ones we had in cdk-assets-schema.